refactor: resume styles
