+++ src/pages/ResumePage/components/AIResumeChecker/index.ts
@@ -1,0 +1,1 @@
+export { AIResumeChecker } from "./AIResumeChecker";
