+++ src/pages/ResumePage/components/ExpandableSection/index.ts
@@ -1,0 +1,1 @@
+export { ExpandableSection } from "./ExpandableSection";
