+++ src/pages/ResumePage/components/HRVideosCarousel/index.ts
@@ -1,0 +1,1 @@
+export { HRVideosCarousel } from "./HRVideosCarousel";
