+++ src/pages/ResumePage/components/MobileResumeGame/MobileResumeGame.tsx
@@ -1,262 +1,23 @@
 import React, { type FC } from "react";
 import { memo, useState } from "react";
-import { 
-  Header, 
-  Text, 
-  Button,
-  mts_brand_red 
-} from "@chernyshovaalexandra/mtsui";
-import type { ResumeOption, GameQuestion } from "../constants";
-import { GAME_QUESTIONS } from "../constants";
-import styled from "styled-components";
-
-const CheckIcon: FC = () => (
-  <svg width="12" height="12" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path 
-      d="M9 16.17l-4.17-4.17-1.42 1.42L9 19 21 7l-1.42-1.42L9 16.17z" 
-      fill="currentColor"
-    />
-  </svg>
-);
-
-const MinusIcon: FC = () => (
-  <svg width="12" height="12" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path 
-      d="M19 13H5v-2h14v2z" 
-      fill="currentColor"
-    />
-  </svg>
-);
-
-const DownloadIcon: FC = () => (
-  <svg  width="18" height="20" viewBox="0 0 18 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path d="M8.99979 0.5C8.37847 0.5 7.87479 1.00368 7.87479 1.625V11.9667C7.75497 11.8159 7.62388 11.6498 7.47934 11.4665C7.02921 10.8957 6.65644 10.4219 6.34873 10.0187C6.1054 9.69983 5.72922 9.50753 5.32813 9.50753C4.39699 9.50753 3.85651 10.4507 4.41663 11.1945C4.77998 11.677 5.2125 12.2255 5.71242 12.8595C6.92588 14.3985 7.53273 15.168 8.28972 15.3946C8.75738 15.5346 9.25578 15.5346 9.72344 15.3946C10.4804 15.168 11.0872 14.3986 12.3006 12.8597C12.8006 12.2256 13.2332 11.6771 13.5965 11.1945C14.1567 10.4507 13.6162 9.50753 12.685 9.50753C12.2839 9.50753 11.9078 9.69983 11.6644 10.0187C11.3567 10.4219 10.984 10.8957 10.5338 11.4665C10.3838 11.6568 10.2483 11.8284 10.1248 11.9837V1.625C10.1248 1.00368 9.62111 0.5 8.99979 0.5Z" fill="currentColor"/>
-    <path d="M1.125 17.7502C0.50368 17.7502 0 18.2539 0 18.8752C0 19.4965 0.50368 20.0002 1.125 20.0002H16.875C17.4963 20.0002 18 19.4965 18 18.8752C18 18.2539 17.4963 17.7502 16.875 17.7502H1.125Z" fill="currentColor"/>
-  </svg>
-);
-
-
-
-const MobileGameContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  padding: 20px;
-  max-width: 100%;
-
-  @media (min-width: 1251px) {
-    display: none;
-  }
-`;
-
-const MobileStepsContainer = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 24px;
-  padding: 0 20px;
-`;
-
-const StepsWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 0;
-`;
-
-const StepCircle = styled.div<{ $isActive: boolean; $isCompleted: boolean }>`
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-weight: 600;
-  font-size: 14px;
-  z-index: 2;
-  position: relative;
-  
-  ${({ $isActive, $isCompleted }) => {
-    if ($isActive) {
-      return `
-        background: ${mts_brand_red};
-        color: white;
-        border: 2px solid ${mts_brand_red};
-      `;
-    } else if ($isCompleted) {
-      return `
-        background: #212529;
-        color: white;
-        border: 2px solid #212529;
-      `;
-    } else {
-      return `
-        background: white;
-        color: #6c757d;
-        border: 2px solid #dee2e6;
-      `;
-    }
-  }}
-
-  .success-indicator,
-  .error-indicator {
-    position: absolute;
-    bottom: -4px;
-    right: -8px;
-    width: 16px;
-    height: 16px;
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    z-index: 3;
-
-    svg {
-      width: 8px;
-      height: 8px;
-    }
-  }
-
-  .success-indicator {
-    background: #26CD58;
-    color: #ffffff;
-  }
-
-  .error-indicator {
-    background: #F95721;
-    color: #ffffff;
-  }
-`;
-
-const StepLine = styled.div<{ $isCompleted: boolean }>`
-  width: 40px;
-  height: 2px;
-  background: ${({ $isCompleted }) => $isCompleted ? '#212529' : '#dee2e6'};
-  transition: background-color 0.3s ease;
-`;
-
-const QuestionTitle = styled(Text)`
-  font-size: 18px;
-  line-height: 140%;
-  color: #212529;
-  font-weight: 500;
-  margin-bottom: 24px;
-  text-align: center;
-`;
-
-const OptionCard = styled.button<{ $isSelected: boolean; $isRevealed: boolean; $isCorrect?: boolean }>`
-  width: 100%;
-  height: 180px;
-  background: ${({ $isSelected, $isCorrect, $isRevealed }) => {
-    if (!$isRevealed) return $isSelected ? '#f8f9fa' : '#e9ecef';
-    if ($isCorrect) return '#d4edda';
-    return $isSelected ? '#f8d7da' : '#e9ecef';
-  }};
-  border: 2px solid ${({ $isSelected, $isCorrect, $isRevealed }) => {
-    if (!$isRevealed) return $isSelected ? mts_brand_red : '#dee2e6';
-    if ($isCorrect) return '#28a745';
-    return $isSelected ? '#dc3545' : '#dee2e6';
-  }};
-  border-radius: 12px;
-  cursor: ${({ $isRevealed }) => $isRevealed ? 'default' : 'pointer'};
-  transition: all 0.2s ease-in-out;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-  margin-bottom: 16px;
-
-  &:hover:not(:disabled) {
-    border-color: ${mts_brand_red};
-    background: #f8f9fa;
-  }
-
-  &:focus-visible {
-    outline: 2px solid ${mts_brand_red};
-    outline-offset: 2px;
-  }
-
-  &:focus:not(:focus-visible) {
-    outline: none;
-  }
-
-  &:disabled {
-    opacity: 0.8;
-  }
-
-  &:last-child {
-    margin-bottom: 0;
-  }
-`;
-
-const PlaceholderText = styled(Text)`
-  color: #6c757d;
-  font-size: 14px;
-  font-weight: 400;
-`;
-
-const ResultBadge = styled.div<{ $isCorrect: boolean }>`
-  position: absolute;
-  top: 12px;
-  right: 12px;
-  background: ${({ $isCorrect }) => $isCorrect ? '#28a745' : '#dc3545'};
-  color: white;
-  padding: 6px 12px;
-  border-radius: 6px;
-  font-size: 12px;
-  font-weight: 600;
-`;
-
-const NavigationButtons = styled.div`
-  display: flex;
-  justify-content: center;
-  gap: 12px;
-  margin-top: 20px;
-`;
-
-const ActionButton = styled(Button)`
-  flex: 1;
-  height: 48px;
-  font-weight: 600;
-`;
-
-const CongratulationsCard = styled.div`
-  background: #ffffff;
-  border-radius: 16px;
-  padding: 24px;
-  text-align: center;
-  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
-  margin-top: 20px;
-`;
-
-const CongratulationsTitle = styled.h3`
-  font-size: 20px;
-  font-weight: 600;
-  color: #212529;
-  margin: 0 0 12px 0;
-`;
-
-const CongratulationsText = styled.p`
-  font-size: 14px;
-  line-height: 140%;
-  color: #6c757d;
-  margin: 0 0 20px 0;
-`;
-
-const DownloadButton = styled(Button)`
-  margin: 0 auto;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-  
-  svg {
-    flex-shrink: 0;
-  }
-`;
-
-
+import { GAME_QUESTIONS } from "../../constants";
+import { CheckIcon, MinusIcon, DownloadIcon } from "../ResumeGame/icons";
+import {
+  MobileGameContainer,
+  MobileStepsContainer,
+  StepsWrapper,
+  StepCircle,
+  StepLine,
+  QuestionTitle,
+  OptionCard,
+  PlaceholderText,
+  NavigationButtons,
+  ActionButton,
+  CongratulationsCard,
+  CongratulationsTitle,
+  CongratulationsText,
+  DownloadButton
+} from "./styles";
 
 interface MobileStepsProps {
   totalSteps: number;
@@ -274,7 +35,6 @@ const MobileSteps: FC<MobileStepsProps> = ({ totalSteps, currentStep, showResult
           const isActive = !showResults && stepNumber === currentStep;
           const isCompleted = stepNumber < currentStep;
           
-          // Показываем результаты для всех шагов только когда showResults=true (в конце игры)
           let stepResult = null;
           if (showResults && selectedAnswers) {
             const question = GAME_QUESTIONS[index];
