+++ src/pages/ResumePage/components/MobileResumeGame/index.ts
@@ -1,0 +1,1 @@
+export { MobileResumeGame } from "./MobileResumeGame";
