+++ src/pages/ResumePage/components/ResumeGame/index.ts
@@ -1,0 +1,1 @@
+export { ResumeGame } from "./ResumeGame";
