+++ src/pages/ResumePage/components/TipsGrid/index.ts
@@ -1,0 +1,1 @@
+export { TipsGrid } from "./TipsGrid";
