+++ src/pages/ResumePage/components/VideoPlayer/index.ts
@@ -1,0 +1,1 @@
+export { VideoPlayer } from "./VideoPlayer";
