+++ src/pages/ResumePage/constants/hrVideos.ts
@@ -14,7 +14,7 @@ export const hrVideos: HRVideo[] = [
     id: "video-1",
     name: "Анна Петрова",
     position: "HR-директор",
-    description: "Расскажу, как составить резюме, которое привлечет внимание рекрутера с первых секунд просмотра.",
+    description: "Расскажу, как составить резюме, которое привлечет внимание рекрутера",
     imageUrl: "/images/hr/anna-petrova.jpg",
     videoUrl: "/videos/hr/anna-petrova.mp4",
     subtitlesUrl: "/videos/hr/anna-petrova.vtt",
@@ -24,7 +24,7 @@ export const hrVideos: HRVideo[] = [
     id: "video-2", 
     name: "Михаил Сидоров",
     position: "Старший рекрутер",
-    description: "Покажу самые распространенные ошибки в резюме и как их избежать на конкретных примерах.",
+    description: "Покажу ошибки в резюме и как их избежать на конкретных примерах.",
     imageUrl: "/images/hr/mikhail-sidorov.jpg",
     videoUrl: "/videos/hr/mikhail-sidorov.mp4",
     subtitlesUrl: "/videos/hr/mikhail-sidorov.vtt",
@@ -34,7 +34,7 @@ export const hrVideos: HRVideo[] = [
     id: "video-3",
     name: "Елена Козлова", 
     position: "IT-рекрутер",
-    description: "Поделюсь особенностями составления резюме для IT-сферы и технических специальностей.",
+    description: "Поделюсь особенностями составления резюме.",
     imageUrl: "/images/hr/elena-kozlova.jpg",
     videoUrl: "/videos/hr/elena-kozlova.mp4",
     subtitlesUrl: "/videos/hr/elena-kozlova.vtt",
@@ -44,7 +44,7 @@ export const hrVideos: HRVideo[] = [
     id: "video-4",
     name: "Дмитрий Волков",
     position: "Ведущий рекрутер",
-    description: "Объясню, как правильно описать навыки и достижения, чтобы выделиться среди других кандидатов.",
+    description: "Объясню, как описать навыки, чтобы выделиться среди других кандидатов.",
     imageUrl: "/images/hr/dmitry-volkov.jpg",
     videoUrl: "/videos/hr/dmitry-volkov.mp4",
     subtitlesUrl: "/videos/hr/dmitry-volkov.vtt",
